Off-diagonal CASPT2 terms via Operator[ IL ][ IR ][ w ][ left + SIZE * right ]
 - IL is the left  center irrep as used in the overlap matrices
 - IR is the right center irrep as used in the overlap matrices
 - w is the active space orbital connecting to the fock operator (of irrep IL x IR)
 - left and right are the standard combined active space indices also used for the overlap matrices

Because there are typically way more doubly occupied / virtual orbitals than active space orbitals,
it will be more efficient to construct intermediates based on "w".

In this form, the Operators can be easily rotated the new basis which removes the overlap matrices
and diagonalizes the diagonal blocks of the CASPT2 Fock operator.